requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# LangGraph and LangChain
langgraph==0.2.16
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0
langchain-core==0.3.0

# Vector Database and Embeddings
faiss-cpu==1.7.4
openai==1.3.0

# Document Processing
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==1.1.0

# Environment and Config
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Additional utilities
aiofiles==23.2.0

# MLflow for experiment tracking
mlflow==2.8.1
scikit-learn==1.3.2

# Database
sqlalchemy==2.0.23
pymysql==1.1.0
alembic==1.13.1